Inputing parameters

+++ models/Hitter.java
@@ -4,12 +4,12 @@ package models;
 public class Hitter extends Player {
     
     //特有屬性
-    short power;                      //力量值
-    short batControl;                 //球棒控制力
-    short plateDiscipline;            //選球
-    short speed;                      //跑壘速度
+    int power;                      //力量值
+    int batControl;                 //球棒控制力
+    int plateDiscipline;            //選球
+    int speed;                      //跑壘速度
     boolean isLefty;                  //是否為左打者
-    protected short[] pitchTypeSights; //球種辨識力
+    protected int[] pitchTypeSights; //球種辨識力
 
     //目前簡化為0:直球 1:變化球
 
@@ -22,8 +22,8 @@ public class Hitter extends Player {
 
     //建構子
     //只有指定打擊會直接使用這個建構子
-    public Hitter(String name, short luck, short collaboration, short clutch, 
-                  short power, short batControl, short plateDiscipline, short speed, boolean isLefty, short[] pitchTypeSights){
+    public Hitter(String name, int luck, int collaboration, int clutch, 
+                  int power, int batControl, int plateDiscipline, int speed, boolean isLefty, int[] pitchTypeSights){
         
         super(name, luck, collaboration, clutch);
         this.speed           = speed;
